+++ defuse/src/intents.rs
@@ -1,5 +1,5 @@
 use defuse_core::{
-    Deadline, Result,
+    Deadline, Result, Salt,
     accounts::{AccountEvent, NonceEvent},
     engine::deltas::InvariantViolated,
     fees::Pips,
@@ -59,6 +59,8 @@ impl SimulationOutput {
 #[derive(Debug, Clone)]
 pub struct StateOutput {
     pub fee: Pips,
+
+    pub current_salt: Salt,
 }
 
 #[ext_contract(ext_relayer_keys)]