+++ tests/src/tests/defuse/accounts/mod.rs
@@ -23,7 +23,11 @@ pub trait AccountManagerExt {
         public_key: PublicKey,
     ) -> anyhow::Result<()>;
 
-    async fn cleanup_nonces(&self, data: &[(AccountId, Vec<Nonce>)]) -> anyhow::Result<TestLog>;
+    async fn cleanup_nonces(
+        &self,
+        defuse_contract_id: &AccountId,
+        data: &[(AccountId, Vec<Nonce>)],
+    ) -> anyhow::Result<TestLog>;
 
     async fn defuse_has_public_key(
         &self,
@@ -87,7 +91,11 @@ impl AccountManagerExt for near_workspaces::Account {
         Ok(())
     }
 
-    async fn cleanup_nonces(&self, data: &[(AccountId, Vec<Nonce>)]) -> anyhow::Result<TestLog> {
+    async fn cleanup_nonces(
+        &self,
+        defuse_contract_id: &AccountId,
+        data: &[(AccountId, Vec<Nonce>)],
+    ) -> anyhow::Result<TestLog> {
         let nonces = data
             .iter()
             .map(|(acc, nonces)| {
@@ -98,7 +106,7 @@ impl AccountManagerExt for near_workspaces::Account {
             .collect::<Vec<(AccountId, Vec<AsBase64<Nonce>>)>>();
 
         let res = self
-            .call(self.id(), "cleanup_nonces")
+            .call(defuse_contract_id, "cleanup_nonces")
             .args_json(json!({
                 "nonces": nonces,
             }))
@@ -195,8 +203,14 @@ impl AccountManagerExt for near_workspaces::Contract {
             .await
     }
 
-    async fn cleanup_nonces(&self, data: &[(AccountId, Vec<Nonce>)]) -> anyhow::Result<TestLog> {
-        self.as_account().cleanup_nonces(data).await
+    async fn cleanup_nonces(
+        &self,
+        defuse_contract_id: &AccountId,
+        data: &[(AccountId, Vec<Nonce>)],
+    ) -> anyhow::Result<TestLog> {
+        self.as_account()
+            .cleanup_nonces(defuse_contract_id, data)
+            .await
     }
 
     async fn defuse_has_public_key(